// server/data/model/bankAccountErrorMessage.ts
export type BankAccountErrorMessage = {
  applicationSubmittedDay?: null | string
  applicationSubmittedMonth?: null | string
  applicationSubmittedYear?: null | string
  isValidDate?: null | string
  validResubmissionDate?: null | string
  validResubmissionHeardBackDate?: null | string
  noBankName?: null | string
  noStatus?: null | string
  dateResubmittedHeardDay?: null | string
  dateResubmittedHeardMonth?: null | string
  dateResubmittedHeardYear?: null | string
  applicationResubmittedDay?: null | string
  applicationResubmittedMonth?: null | string
  applicationResubmittedYear?: null | string
  accountOpenedDay?: null | string
  accountOpenedMonth?: null | string
  accountOpenedYear?: null | string
  validAccountOpenedDate?: null | string
  dateAddedDay?: null | string
  dateAddedMonth?: null | string
  dateAddedYear?: null | string
  validDateAdded?: null | string
  heardBackDay?: null | string
  heardBackMonth?: null | string
  heardBackYear?: null | string
  validHeardBackDate?: null | string
  noPersonalItems?: null | string
}